+++ app/Http/Controllers/YearsController.php
@@ -4,33 +4,28 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Tahun;
+use Illuminate\Support\Facades\Auth;
+
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class YearsController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     */
+    use AuthorizesRequests;
     public function index(Request $request)
     {
-        $query=Tahun::query()->where('user_id',auth()->id());
+        $this->authorize('viewAny',Tahun::class);
+        $query=Tahun::query()->where('user_id',Auth::id());
         if($request->has('search') && $request->search != ''){
             $query->where('tahun','like','%'.$request->search.'%');
         }
         $years=$query->paginate(10);
         return view('years.index',compact('years'));
     }
-
-    /**
-     * Show the form for creating a new resource.
-     */
     public function create()
     {
+        $this->authorize('create',Tahun::class);
         return view('years.create');
     }
-
-    /**
-     * Store a newly created resource in storage.
-     */
     public function store(Request $request)
     {
         $request->validate([
@@ -38,48 +33,37 @@ class YearsController extends Controller
         ]);
         $years=Tahun::create([
             'tahun'=>$request->tahun,
-            'user_id'=>auth()->id(),
+            'user_id'=>Auth::id(),
         ]);
+        $this->authorize($years,Tahun::class);
         return redirect()->route('create.years')->with('success','Berhasil menambah tahun rilis');
     }
-
-    /**
-     * Display the specified resource.
-     */
     public function show(string $id)
     {
         $years=Tahun::findOrFail($id);
+        $this->authorize($years,Tahun::class);
         return view('years.show',compact('years'));
     }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
     public function edit(string $id)
     {
         $years=Tahun::findOrFail($id);
+        $this->authorize($years,Tahun::class);
         return view('years.edit',compact('years'));
     }
-
-    /**
-     * Update the specified resource in storage.
-     */
     public function update(Request $request, string $id)
     {
         $request->validate([
             'tahun'=>'required',
         ]);
         $years=Tahun::findOrFail($id);
+        $this->authorize($years,Tahun::class);
         $years->update($request->all());
         return redirect()->route('daftar.years')->with('success','Berhasil mengubah tahun rilis');
     }
-
-    /**
-     * Remove the specified resource from storage.
-     */
     public function destroy(string $id)
     {
         $years=Tahun::findOrFail($id);
+        $this->authorize($years,Tahun::class);
         $years->delete();
         return redirect()->route('daftar.years')->with('success','Berhasil menghapus');
     }
